#1991 Split ModuleManager and ModuleInformation into two files.
