+++ src/psyclone/parse/module_manager.py
@@ -39,83 +39,10 @@ which module is contained in which file (including full location). '''
 
 import os
 
-from fparser.common.readfortran import FortranStringReader
-from fparser.two.Fortran2003 import Use_Stmt
-from fparser.two.parser import ParserFactory
-from fparser.two.utils import walk
-
 from psyclone.errors import InternalError
+from psyclone.parse.module_info import ModuleInfo
 
 
-# ============================================================================
-class ModuleInfo:
-    '''This class stores mostly cached information about modules: it stores
-    the original filename, if requested it will read the file and then caches
-    the plain text file, and if required it will parse the file, and then
-    cache the fparser AST.
-
-    :param str name: the module name.
-    :param str filename: the name of the source file that stores this module \
-        (including path).
-
-    '''
-
-    def __init__(self, name, filename):
-        self._name = name
-        self._filename = filename
-        # A cache for the source code:
-        self._source_code = None
-        # A cache for the fparser tree
-        self._parse_tree = None
-
-    # ------------------------------------------------------------------------
-    @property
-    def filename(self):
-        ''':returns: the filename that contains the source code for this
-        module.
-        :rtype: str
-
-        '''
-        return self._filename
-
-    # ------------------------------------------------------------------------
-    def get_source_code(self):
-        '''Returns the source code for the module. The first time, it
-        will be read from the file, but the data is then cached.
-
-        :returns: the source code as string.
-        :rtype: str
-
-        :raises FileNotFoundError: when the file cannot be read.
-
-        '''
-        if self._source_code is None:
-            try:
-                with open(self._filename, "r", encoding='utf-8') as file_in:
-                    self._source_code = file_in.read()
-            except FileNotFoundError as err:
-                raise FileNotFoundError(
-                    f"Could not find file '{self._filename}' when trying to "
-                    f"read source code for module '{self._name}'") from err
-
-        return self._source_code
-
-    # ------------------------------------------------------------------------
-    def get_parse_tree(self):
-        '''Returns the fparser AST for this module. The first time, the file
-        will be parsed by fpaser, then the AST is cached for any future uses.
-        :returns: the fparser AST for this module.
-        :rtype:
-
-        '''
-        if not self._parse_tree:
-            reader = FortranStringReader(self.get_source_code())
-            parser = ParserFactory().create(std="f2003")
-            self._parse_tree = parser(reader)
-        return self._parse_tree
-
-
-# ============================================================================
 class ModuleManager:
     '''This class implements a singleton that manages module
     dependencies.
@@ -257,52 +184,3 @@ class ModuleManager:
             return [root]
 
         return []
-
-    # ------------------------------------------------------------------------
-    def find_modules_used_in(self, module_name):
-        '''This function analyses a given module source file and returns
-        a list of 2-tuples containing the module name, and a list of
-        all imported symbols from that module. If all symbols are imported,
-        the list of symbols will be empty.
-
-        :param str module_name: the file name (including path if required) \
-            for which all modules should be found.
-
-        :returns: a list of with all module names and symbols imported.
-        :rtype: List[Tuple[str, List[str]]]
-
-        '''
-        module_name = module_name.lower()
-
-        # Check if we have cached results:
-        result = self._cached_module_info.get(module_name, None)
-        if result is not None:
-            return result
-
-        # Otherwise, parse the file:
-        try:
-            mod_info = self.get_module_info(module_name)
-        except FileNotFoundError:
-            return []
-
-        parse_tree = mod_info.get_parse_tree()
-        results = []
-        for use in walk(parse_tree, Use_Stmt):
-            # Ignore intrinsic modules:
-            if str(use.items[0]) == "INTRINSIC":
-                continue
-
-            only_list = use.items[4]
-            if not only_list:
-                # Imports everything. Indicate this by
-                # using an empty list for the symbols
-                results.append((str(use.items[2]), []))
-                continue
-            # Parse the only list:
-            all_symbols = []
-            for symbol in use.items[4].children:
-                all_symbols.append(str(symbol))
-            results.append((str(use.items[2]), all_symbols))
-
-        self._cached_module_info[module_name] = results
-        return results
